+++ test/vident/component_test.rb
@@ -7,6 +7,8 @@ module Vident
       @test_component_class = Class.new do
         include Vident::Component
 
+        prop :my_string, _Nilable(String)
+
         def self.name
           "TestComponent"
         end
@@ -254,6 +256,12 @@ module Vident
         id: "test-id"
       }
       assert_equal expected, options
+    end
+
+    def test_prop_names
+      all_props = [:stimulus_controllers, :stimulus_actions, :stimulus_targets, :stimulus_outlets, :stimulus_outlet_host, :stimulus_values, :stimulus_classes, :element_tag, :id, :classes, :html_options, :my_string]
+      assert_equal all_props, @test_component_class.prop_names
+      assert_equal all_props, @test_component_class.new.prop_names
     end
   end
 end